Added RevisionCache interface, and renamed existing implementations to be LRU specific
